Miguel prueba Popup de felipe

+++ src/main/java/com/verygana2/dtos/raffle/websocket/WaitingRoomPayloadDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class WaitingRoomPayloadDTO {
 
-    private int viewerCount; // Usuarios conectados a esta sala
+    private long viewerCount; // Usuarios conectados a esta sala
     private long secondsUntilDraw; // Para el countdown del frontend
     private long totalTickets; // totalTicketsIssued de la rifa
     private Long totalParticipants;
